Performance improvement for issue #60. Add NO_TOPSORT to avoid topological sorting (flattening of graph)
Issue is currently that the topological sorting is done too often inside
the EDA backend functions. That makes the configure time explode
exponentionally.
There is no good reason to flatten the graph unless ip_link() was called
on a node in the tree. Since it is not easy to detect that it was
called, we cannot safely say when we need to call flatten_graph() when
we dont.
Since the EDA backed functions are controlled, we can perform
topological sorting at the beginning and then use get_ip_...(NO_TOPSORT)
to ensure each get_ip_...() call is not performing a redundant
topological sorting